requirements.txt
# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
torch>=1.12.0
gymnasium>=0.28.0

# MuJoCo
mujoco>=2.3.0

# RL algorithms
stable-baselines3>=2.0.0
sb3-contrib>=2.0.0

# Visualization and analysis
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.62.0
rich>=12.0.0
typer>=0.6.0

# Logging and monitoring
tensorboard>=2.8.0
wandb>=0.13.0

# Hyperparameter optimization
optuna>=3.0.0

# Configuration management
hydra-core>=1.2.0
omegaconf>=2.2.0

# Development tools
jupyter>=1.0.0
ipykernel>=6.0.0
